requirements.txt
office365-rest-python-client
openpyxl
tkcalendar
python-environ
requests
pandas
kivy
kivy-garden
kivymd